(1)  Fix problem loading message files with DOS EOL characters (2)  Fix drawing problem for the money string (3)  Add missing function prototypes
git-svn-id: svn://svn.berlios.de/lincity-ng/trunk@137 8226ee30-1aee-0310-a633-b68f58d1eb45

+++ lintypes.c
@@ -1517,5 +1517,5 @@ get_type_name (short type, char * s)
 int
 get_type_cost (short type)
 {
-    return get_group_cost (get_group_of_type (type));
+    return get_group_cost ((short) get_group_of_type (type));
 }
